refactored testcases and setup to there own modules
